exception improvment

+++ src/com/internectics/data/Answer.java
@@ -123,12 +123,12 @@ public class Answer {
         String query = String.format("DELETE FROM Answer_Tables WHERE card_id=%d", cardID);
         SQLiteHelper.defaultDatabase(context).execSQL(query);
 
-        if (!StringUtils.isNumeric(imageUriFormatStr)) {
+        if (!StringUtils.isNumeric(imageUriFormatStr) && (!imageUriFormatStr.contains("answer_placeholder_content.jpg"))) {
             File file = new File(FileOperationHelper.deleteUriSchemeHeader(this.imageUriFormatStr));
             if (file.delete()) {
-                Log.d(Global.debugTag, "Successful to delete imageUriFormatStr file in Answer");
+                //Log.d(Global.debugTag, "Successful to delete imageUriFormatStr file in Answer");
             } else {
-                Log.e(Global.debugTag, "Fail to delete coverImageUriFormatStr file in Answer");
+                Log.e(Global.debugTag, "Fail to delete coverImageUriFormatStr file in Answer:" + file);
             }
         }
     }
